modified the toy controller

+++ app/controllers/toys_controller.rb
@@ -14,12 +14,17 @@ class ToysController < ApplicationController
   def update
     toy = Toy.find_by(id: params[:id])
     toy.update(toy_params)
+    render json: toy, status: :accepted
   end
 
   def destroy
     toy = Toy.find_by(id: params[:id])
-    toy.destroy
-    head :no_content
+    if toy
+      toy.destroy
+      head :no_content
+    else
+      render json: toy.errors.full_messages, status: :not_found
+    end
   end
 
   private
@@ -27,5 +32,6 @@ class ToysController < ApplicationController
   def toy_params
     params.permit(:name, :image, :likes)
   end
+
 
 end
